+++ x/leverage/keeper/grpc_query_test.go
@@ -5,6 +5,8 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	"github.com/umee-network/umee/v3/x/leverage/fixtures"
+	"github.com/umee-network/umee/v3/x/leverage/keeper"
 	"github.com/umee-network/umee/v3/x/leverage/types"
 )
 
@@ -21,7 +23,7 @@ func (s *IntegrationTestSuite) TestQuerier_Params() {
 
 	resp, err := s.queryClient.Params(ctx.Context(), &types.QueryParams{})
 	require.NoError(err)
-	require.Equal(types.DefaultParams(), resp.Params)
+	require.Equal(fixtures.Params(), resp.Params)
 }
 
 func (s *IntegrationTestSuite) TestQuerier_MarketSummary() {
@@ -117,6 +119,13 @@ func (s *IntegrationTestSuite) TestQuerier_AccountSummary() {
 
 func (s *IntegrationTestSuite) TestQuerier_LiquidationTargets() {
 	ctx, require := s.ctx, s.Require()
+
+	keeper.EnableLiquidator = "false"
+
+	_, err := s.queryClient.LiquidationTargets(ctx.Context(), &types.QueryLiquidationTargets{})
+	require.ErrorIs(err, types.ErrNotLiquidatorNode)
+
+	keeper.EnableLiquidator = "true"
 
 	resp, err := s.queryClient.LiquidationTargets(ctx.Context(), &types.QueryLiquidationTargets{})
 	require.NoError(err)