+++ price-feeder/CHANGELOG.md
@@ -46,9 +46,16 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Bugs
+
+- [1428](https://github.com/umee-network/umee/pull/1428) Update umeed version to an actual tag.
+
+## [v1.0.0](https://github.com/umee-network/umee/releases/tag/price-feeder%2Fv1.0.0) - 2022-09-19
+
 ### Features
 
 - [1328](https://github.com/umee-network/umee/pull/1328) Add bitget provider.
+- [1339](https://github.com/umee-network/umee/pull/1339) Add mexc provider.
 
 ### Bugs